♻️ refactor: constants.js 네이밍 수정 및 TMDB_MOVIE_API_BASE_URL 구조 단순화
