+++ src/constants.js
@@ -1,12 +1,10 @@
-export const TMDB_API_BASE_URL = "https://api.themoviedb.org/3";
-
-export const TMDB_MOVIE_API_BASE_URL = `${TMDB_API_BASE_URL}/movie`;
+export const TMDB_MOVIE_API_BASE_URL = "https://api.themoviedb.org/3/movie";
 
 export const TMDB_GET_OPTION = {
   method: "GET",
   headers: {
     accept: "application/json",
-    Authorization: `Bearer ${import.meta.env.VITE_TMDP_READ_ACCESS_TOKEN}`,
+    Authorization: `Bearer ${import.meta.env.VITE_TMDB_READ_ACCESS_TOKEN}`,
   },
 };
